Remove TripDetailsClient and page.jsx components; refactor to streamline trip details handling and improve server-side rendering.
